resolve suggestions
- move FormatTypeParams from internal/golang/types_format.go to
  internal/util/typesutil.go
- resolve TODO that check if the type is function-local early
